+++ app/src/Components/Committee/CommitteeList/CommitteeCard/CommitteeCard.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import classes from './CommitteeCard.module.css';
+
+const CommitteeCard = (props) => {
+    
+}
+
+export default CommitteCard;
