+++ app/src/Components/Document/Document.js
@@ -21,7 +21,7 @@ class Document extends Component {
 
         this.state = {
             viewCardsview: true,
-            propUserFunkis: false// Ska komma som prop
+            propUserFunkis: true// Ska komma som prop
         }
     }
